1 Corrigindo nomes de variáveis que não fazia sentido (retirando várias referências a tuples, quando podiam não ser de fato tuples) 2 Corrigindo cast que estava para tuple no CompareSimpleRegister 3 Possibilitando o update para o listener de Folders

+++ project/datawatcher/src/main/java/br/com/datawatcher/entity/Tuple.java
@@ -60,7 +60,7 @@ public class Tuple extends SimpleRegister {
 	}
 	
 	@Override
-	public int hashSimpleRegister() {
+	public long hashSimpleRegister() {
 		if (this.hashRegister == null) {
 			this.calculatesHashes();
 		}
